+++ app/components/Views/MultichainAccounts/2accountDetails/2accountDetails.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
 import { InternalAccount } from '@metamask/keyring-internal-api';
-import { BaseAccountDetails } from './AccountTypes/BaseAccountDetails';
+import { BaseAccountDetails } from './2accountTypes/2baseAccountDetails';
 
 interface AccountDetailsProps {
   route: {
