+++ app/components/Views/MultichainAccounts/2accountDetails/2accountTypes/2baseAccountDetails/styles.ts
