+++ app/components/Views/MultichainAccounts/2accountDetails/components/2accountInfo/2accountInfo.styles.ts
