+++ app/components/Views/MultichainAccounts/2accountDetails/components/2accountInfo/2accountInfo.tsx
@@ -10,7 +10,7 @@ import Avatar, {
 } from '../../../../../../component-library/components/Avatars/Avatar';
 import { InternalAccount } from '@metamask/keyring-internal-api';
 import { useStyles } from '../../../../../hooks/useStyles';
-import styleSheet from './AccountInfo.styles';
+import styleSheet from './2accountInfo.styles';
 import Text, {
   TextVariant,
 } from '../../../../../../component-library/components/Texts/Text';
