+++ app/components/Views/MultichainAccounts/2accountDetails/components/2accountInfo/index.ts
@@ -1,0 +1,1 @@
+export { AccountInfo as default } from './2accountInfo';
