--- app/components/Views/MultichainAccounts/accountDetails/components/accountInfo/index.ts
@@ -1,1 +1,0 @@
-export { AccountInfo as default } from './AccountInfo';
